separate client and server side components
